+++ dyndis/__data__.py
@@ -1,2 +1,3 @@
 __author__ = 'ben avrahami'
-__version__ = '0.0.1'
+__version__ = '0.0.2.dev'
+__url__ = "https://github.com/bentheiii/dyndis"