+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# leyline changelog
+# dyndis changelog
+
+## 0.0.2: unreleased
+### added
+* trie improved candidate lookup
+### removed
+* got rid of the priority alias, just number now
 
 ## 0.0.1: 2019-10-28
 * initial